move WebPage to separate file
patch eslint and nock
