+++ package.json
@@ -36,11 +36,11 @@
   "devDependencies": {
     "@bonniernews/wichita": "^1.1.0",
     "chai": "^4.3.6",
-    "eslint": "^8.14.0",
+    "eslint": "^8.23.0",
     "express": "^4.18.1",
     "markdown-toc": "^1.2.0",
     "mocha": "^10.0.0",
-    "nock": "^13.2.4"
+    "nock": "^13.2.9"
   },
   "files": [
     "lib/",
